pyproject.toml
[project]
name = "xrc-autocot-experiment"
version = "0.1.0"
description = "Cross-Reasoning Consistency Auto-CoT experiment"
requires-python = ">=3.11"

dependencies = [
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "openai>=1.12.0",
    "datasets>=2.16.0",
    "scikit-learn>=1.4.0",
    "sentence-transformers>=2.3.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]